+++ packages/core/src/canvas/index.ts
@@ -6,6 +6,7 @@ export * from "./renderLink";
 export * from "./renderLinkRadial";
 export * from "./renderPolygon";
 export * from "./renderRect";
+export * from "./renderRibbon";
 export * from "./renderText";
 export * from "./renderElements";
 export * from "./virtual";